Sample Play+Bootstrap+Markdown app for asakai
A very simple slideshow application built with Play.

Slides are written in Markdown and saved as "<slideId>.md" in the
slides/ directory.

The controller reads these Markdown files, extracts the slide title and
content, converts it to HTML using Pegdown, and sends it off to a
template for rendering.

Note that slide titles must be on the first line, and must use the
underline (Setext) syntax rather than the hash (atx) syntax. So the first
2 lines of the markdown file should look like:

MyTitle
=======

There is also an index of all the slides. Here the controller loops
through all .md files in the slides dir, extracts the slide ID (from the
filename) and the title (the first line of the file) and sends them off
to a template for rendering in a table.